move common helpers to lib/helpers/
