+++ lib/models/competition-events/svgf-rh.freestyle@2020.ts
@@ -1,7 +1,7 @@
 import { RSRWrongJudgeTypeError } from '../../errors'
-import { clampNumber, filterMarkStream, filterTally, matchMeta, roundTo, roundToCurry, simpleCalculateTallyFactory } from '../../helpers'
+import { clampNumber, filterMarkStream, filterTally, matchMeta, roundTo, roundToCurry, simpleCalculateTallyFactory } from '../../helpers/helpers'
 import { type ScoreTally, type JudgeTypeGetter, type TableDefinition, type CompetitionEventModel } from '../types'
-import { ijruAverage } from './ijru.common'
+import { ijruAverage } from '../../helpers/ijru'
 
 type Option = 'discipline'
